+++ npm/esbuild/install.js
@@ -1,9 +1,9 @@
 const fs = require('fs');
 const os = require('os');
 const path = require('path');
-const child_process = require('child_process');
+const zlib = require('zlib');
+const https = require('https');
 const version = require('./package.json').version;
-const installDir = path.join(__dirname, '.install');
 const binPath = path.join(__dirname, 'bin', 'esbuild');
 const stampPath = path.join(__dirname, 'stamp.txt');
 
@@ -16,51 +16,39 @@ function installBinaryFromPackage(package, fromPath, toPath) {
     return;
   }
 
-  // Clone the environment without "npm_" environment variables. If we don't do
-  // this, invoking this script via "npm install -g esbuild" will hang because
-  // our call to "npm install" below will magically be transformed into
-  // "npm install -g" and, I assume, deadlock waiting for the global lock.
-  const env = {};
-  for (const key in process.env) {
-    if (!key.startsWith('npm_')) {
-      env[key] = process.env[key];
-    }
-  }
+  // Download the package from npm
+  let url = `https://registry.npmjs.org/${package}/-/${package}-${version}.tgz`
+  downloadURL(url, buffer => {
+    // Extract the binary executable from the package
+    fs.writeFileSync(toPath, extractFileFromTarGzip(buffer, fromPath));
 
-  // Run "npm install" recursively to install this specific package
-  fs.mkdirSync(installDir);
-  fs.writeFileSync(path.join(installDir, 'package.json'), '{}');
-  child_process.execSync(`npm install --silent --prefer-offline --no-audit --progress=false ${package}@${version}`,
-    { cwd: installDir, stdio: 'inherit', env });
-
-  // Move the binary from the nested package into our package
-  fs.renameSync(fromPath, toPath);
-
-  // Remove the install directory afterwards to avoid tripping up tools that scan
-  // for nested directories named "node_modules" and make assumptions. See this
-  // issue for an example: https://github.com/ds300/patch-package/issues/243
-  removeRecursive(installDir);
-
-  // Mark the operation as successful so this script is idempotent
-  fs.writeFileSync(stampPath, '');
+    // Mark the operation as successful so this script is idempotent
+    fs.writeFileSync(stampPath, '');
+  });
 }
 
-function removeRecursive(dir) {
-  for (const entry of fs.readdirSync(dir)) {
-    const entryPath = path.join(dir, entry);
-    let stats;
-    try {
-      stats = fs.lstatSync(entryPath);
-    } catch (e) {
-      continue; // Guard against https://github.com/nodejs/node/issues/4760
-    }
-    if (stats.isDirectory()) {
-      removeRecursive(entryPath);
-    } else {
-      fs.unlinkSync(entryPath);
+function downloadURL(url, done) {
+  https.get(url, res => {
+    let chunks = [];
+    res.on('data', chunk => chunks.push(chunk));
+    res.on('end', () => done(Buffer.concat(chunks)));
+  }).on('error', err => { throw err; });
+}
+
+function extractFileFromTarGzip(buffer, file) {
+  buffer = zlib.unzipSync(buffer);
+  let str = (i, n) => String.fromCharCode(...buffer.subarray(i, i + n)).replace(/\0.*$/, '');
+  let offset = 0;
+  while (offset < buffer.length) {
+    let name = str(offset, 100);
+    let size = parseInt(str(offset + 124, 12), 8);
+    offset += 512;
+    if (!isNaN(size)) {
+      if (name === file) return buffer.subarray(offset, offset + size);
+      offset += (size + 511) & ~511;
     }
   }
-  fs.rmdirSync(dir);
+  throw new Error(`Could not find ${JSON.stringify(file)}`)
 }
 
 function installOnUnix(package) {
@@ -68,25 +56,11 @@ function installOnUnix(package) {
     fs.unlinkSync(binPath);
     fs.symlinkSync(process.env.ESBUILD_BIN_PATH_FOR_TESTS, binPath);
   } else {
-    installBinaryFromPackage(
-      package,
-      path.join(installDir, 'node_modules', package, 'bin', 'esbuild'),
-      binPath
-    );
+    installBinaryFromPackage(package, 'package/bin/esbuild', binPath);
   }
 }
 
 function installOnWindows() {
-  const exePath = path.join(__dirname, 'esbuild.exe');
-  if (process.env.ESBUILD_BIN_PATH_FOR_TESTS) {
-    fs.symlinkSync(process.env.ESBUILD_BIN_PATH_FOR_TESTS, exePath);
-  } else {
-    installBinaryFromPackage(
-      'esbuild-windows-64',
-      path.join(installDir, 'node_modules', 'esbuild-windows-64', 'esbuild.exe'),
-      exePath
-    );
-  }
   fs.writeFileSync(
     binPath,
     `#!/usr/bin/env node
@@ -94,8 +68,13 @@ const path = require('path');
 const esbuild_exe = path.join(__dirname, '..', 'esbuild.exe');
 const child_process = require('child_process');
 child_process.spawnSync(esbuild_exe, process.argv.slice(2), { stdio: 'inherit' });
-`
-  );
+`);
+  const exePath = path.join(__dirname, 'esbuild.exe');
+  if (process.env.ESBUILD_BIN_PATH_FOR_TESTS) {
+    fs.symlinkSync(process.env.ESBUILD_BIN_PATH_FOR_TESTS, exePath);
+  } else {
+    installBinaryFromPackage('esbuild-windows-64', 'package/esbuild.exe', exePath);
+  }
 }
 
 const knownUnixlikePackages = {